requirements.txt
streamlit==1.31.1
opencv-python-headless==4.8.0.76
ultralytics==8.0.196
Pillow==10.0.1
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.18.5
pyyaml>=5.3.1
tqdm>=4.64.0
matplotlib>=3.3.0
seaborn>=0.11.0